marketplace/apis/registration/requirements.txt
fastapi
uvicorn
pydantic[email]
pika
python-dotenv